requirements.txt
playwright==1.49.1
ddddocr==1.4.11
openpyxl==3.1.5
Pillow==11.1.0
loguru==0.7.3
